feat: Add routing
Add routing
Move code from App.tsx to separate pages file (Notes.tsx)
